test(input): migrate bottom-content
