+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-bottom-content-counter-ios-rtl-Mobile-Chrome-linux.png
