+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-bottom-content-counter-md-ltr-Mobile-Safari-linux.png
