+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-bottom-content-counter-md-rtl-Mobile-Chrome-linux.png
