+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-bottom-content-error-ios-ltr-Mobile-Chrome-linux.png
