+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-bottom-content-helper-md-ltr-Mobile-Firefox-linux.png
