+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-bottom-content-helper-md-rtl-Mobile-Safari-linux.png
