+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-error-custom-color-md-ltr-Mobile-Safari-linux.png
