+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-full-bottom-no-fill-ios-ltr-Mobile-Safari-linux.png
