+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-full-bottom-outline-ios-ltr-Mobile-Firefox-linux.png
