+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-full-bottom-outline-md-ltr-Mobile-Firefox-linux.png
