+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-full-bottom-solid-ios-ltr-Mobile-Chrome-linux.png
