+++ core/src/components/input/test/bottom-content/input.e2e.ts-snapshots/input-full-bottom-solid-md-ltr-Mobile-Chrome-linux.png
